Moved TopBar to its own file
