+++ src/components/index.ts
@@ -3,5 +3,8 @@ import { NewPost } from "./NewPost";
 import { UserData } from "./UserData";
 import { People } from "./People";
 import { StyledFeed as Feed } from "./Feed";
+import { Login } from "./Login";
+import { TopBar } from "./TopBar";
+import { Register } from "./Register";
 
-export { Post, NewPost, UserData, People, Feed };
+export { Post, NewPost, UserData, People, Feed, Login, TopBar, Register };
